requirements.txt
discord.py==2.3.2
tweepy==4.14.0
gspread==5.12.0
google-auth==2.23.4
pandas==2.1.1
gspread-dataframe==3.3.1
python-dotenv==1.0.0
nltk==3.8.1